requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.1.3
matplotlib==3.6.2
seaborn==0.12.2
jupyter==1.0.0
pytest==7.2.0
pyyaml==6.0
joblib==1.2.0
lightgbm==3.3.2